Reorganize Platform folder.
