+++ src/Platforms/CUDA_legacy/CMakeLists.txt
@@ -17,6 +17,5 @@ SET(CUDA_LEGACY_SRCS
 )
 
 CUDA_ADD_LIBRARY(platform_cuda_legacy ${CUDA_LEGACY_SRCS})
-TARGET_INCLUDE_DIRECTORIES(platform_cuda_legacy INTERFACE "${CMAKE_CURRENT_SOURCE_DIR}")
 CUDA_ADD_CUBLAS_TO_TARGET(platform_cuda_legacy)
 TARGET_LINK_LIBRARIES(platform_cuda_legacy PRIVATE ${CUDA_LIBRARIES})
