+++ src/Platforms/OpenMP/CMakeLists.txt
@@ -19,11 +19,7 @@ IF(USE_OBJECT_TARGET)
 ELSE()
   ADD_LIBRARY(platform_omp ${OMP_SRCS})
 ENDIF()
-TARGET_INCLUDE_DIRECTORIES(platform_omp INTERFACE "${CMAKE_CURRENT_SOURCE_DIR}")
+
 IF(NOT QMC_COMPLEX)
   TARGET_COMPILE_DEFINITIONS(platform_omp PRIVATE -DOPENMP_NO_COMPLEX)
-ENDIF()
-
-IF (BUILD_UNIT_TESTS)
-  SUBDIRS(tests)
 ENDIF()
