+++ src/Platforms/tests/CUDA/CMakeLists.txt
@@ -2,7 +2,7 @@
 #// This file is distributed under the University of Illinois/NCSA Open Source License.
 #// See LICENSE file in top directory for details.
 #//
-#// Copyright (c) 2020 QMCPACK developers.
+#// Copyright (c) 2019 QMCPACK developers.
 #//
 #// File developed by: Ye Luo, yeluo@anl.gov, Argonne National Laboratory
 #//
@@ -10,9 +10,13 @@
 #//////////////////////////////////////////////////////////////////////////////////////
 
 
-SET(CPU_SRCS
-  BlasThreadingEnv.cpp
-)
+SET(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${QMCPACK_UNIT_TEST_DIR})
 
-ADD_LIBRARY(platform_cpu ${CPU_SRCS})
-TARGET_LINK_LIBRARIES(platform_cpu Math::BLAS_LAPACK Math::scalar_vector_functions)
+SET(SRC_DIR CUDA)
+SET(UTEST_EXE test_${SRC_DIR})
+SET(UTEST_NAME deterministic-unit_test_${SRC_DIR})
+
+ADD_EXECUTABLE(${UTEST_EXE} test_CUDAallocator.cpp)
+TARGET_LINK_LIBRARIES(${UTEST_EXE} platform_device containers catch_main)
+
+ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
