+++ src/Platforms/tests/OpenMP/CMakeLists.txt
@@ -19,7 +19,7 @@ SET(UTEST_EXE test_${SRC_DIR})
 SET(UTEST_NAME deterministic-unit_test_${SRC_DIR})
 
 ADD_EXECUTABLE(${UTEST_EXE} test_vector.cpp test_math.cpp test_deep_copy.cpp test_class_member.cpp)
-TARGET_LINK_LIBRARIES(${UTEST_EXE} platform_omp catch_main)
+TARGET_LINK_LIBRARIES(${UTEST_EXE} platform_device platform_omp catch_main)
 
 ADD_UNIT_TEST(${UTEST_NAME} "${QMCPACK_UNIT_TEST_DIR}/${UTEST_EXE}")
