+++ src/QMCApp/accelerators.hpp
